+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes
 
-## v0.1.0 - Unreleased
+## v0.1.1 - Unreleased
+
+- Status bar menu now shows all register mode commands, not just the commands
+  from the currently active mode.
+- Improvements to build/test system
+
+## v0.1.0 - 2026/09/17
 
 - Initial release.